1733 has the "alpha"-level implementation of the new copy paradigm
For now, this enables, in principle, any kind of object to be copied
to any other kind of object. It is now possible to copy a different
Matrix type, ex. int to float. We will expand this a bit to 
allow copying of DataCol and Tables where you leave your cols
in place, and just copy the data. We will probably also find
use for this in specs, networks, and such.

A number of "core" changes have been made, which may break things.

Names:

 * we no longer copy the object name by default -- this may break
   some things, ex it broke DataCol, so I had to add that in
   (for now)
   
Lists:

 * duplicate now puts the el after the one you choose, and gives
   it the default name