A little benchmark tool, that benchmarks the performance of the Allreduce function.
Given, that the Allreduce function appears to be the bottleneck for scaling emergent to larger n with MPI, 
its performance is very critical. On the otherhand, on default settings, the standard MPI_Allreduce performance 
appears to be very problematic on a number of systems. So this lets one determine the performance under similar 
conditions as would be experienced in emergent. 
 
This also reimplements the Allreduce function in a separate class to compare a "naive" implementation to the 
standard MPI libraries. Given on a number of systems, this actually performs considerably better than the 
MPI libraries, it allows to identify how far off the tuning of the MPI libraries are from the achievable 
performance on the system.