potential fix to locale issue: just force the damn thing to always use the "C" locale -- given that everything is based strongly on C++ code in the program system, it is not really possible to support non-C numbers anyway (a comma means something different than a dot in C!), so we should just go whole hog.
I briefly considered using the QString::number routines in taString but they take a
different formatting arg and it would be a fair bit of work and incompatibiltity to
switch over to using that instead.  In any case, the relevant ta_snprintf routine
is now within the .cpp file if we decide to do this at some point.

But I'm pretty sure this is the right fix.. John can you test it??