call taMisc::DoneBusy in the ClusterManager update routines in the exception handling
In ClusterManager::UpdateWorkingCopy_impl Busy() is called around the svn commands. However,
there is no try catch at that level and it falls through to the catch clause in ClusterManager::UpdateTables()
Add a DoneBusy() there.

Need to double check the other possible call stack paths to call ClusterManager::UpdateWorkingCopy to see if we need to do the same
in those catch clauses.