Try and pass through errors to emergent that happen during job submission.
This is still not ideal, because now if you submit a job and it failes to reach the job scheduer,
e.g. because one gave the wrong queue, or because the path variables are set wrong and it can't find pyqsub,
the job directly goes to the jobs_done table and never shows up in the job running.

As one is expecting the job to show up in the jobs_running table and not jobs_done table, it is quite likely
that someone still misses the fact that the job never reached the scheduler.