Add a "GetMember" command to the server protocol.
This should allow to retrieve an arbitrary member from the object tree via the server protocol.

The json command over the server protocol would be something like:
{"command": "GetProperty", "path": ".projects[\"bvPVLV\"].networks[\"bvPVLVNet\"].specs.gp[\"PVLV\"][\"PVLVLayers\"].children[\"AmygLayer\"].children[\"AmygExtLayer\"].clamp", "member": "gain"}
to retrieve the gain member from one of the layer specs.

This command should be useful for some of the testing to not need to each time write a emergent program to extract a member into a ProgVar and then use "GetVar".

It could also be useful in case one wants to create python bindings to emergent over the server protocol.

This code probably still needs some cleanup to adhere to the coding conventions or other code hygene.