Catch SVN exception when trying to create new repository.
If the cluster/username directory already exists in the svn repository, but
we didn't have a working copy checked out, emergent tried to create the directory
again in the svn repository which failed. But emergent didn't handle the case.

As we are doing a checkout after the make anyway, simply ignore the svn error on make.
If the make failed because it already exists, then great, otherwise if it failed for a
different reason, the subsequent svn checkout should fail and emergent processes that error
in a similar way as it would have previously handled the make dir exception