Add a simple test for loading and saving weights.
This Runs a simple network for a while, then saves the weights, loads the weights and then saves the weights again.
The test then compares the two saved weight files to make sure the diff is empty.

This can only find cases when loading and saving weights are inconsistent, but not if they are consistent but broken.