Add a Hippocampus wizard test
This test calls the hippo wizard on a network and then runs MasterTrain on the standard (non hippocampus specific) input data.

At the moment, this doesn't yet actually test any of the hippocampus functionality, and the wizard doesn't even connect
the new hippo layers into the standard layers. So at the moment, this just tests that it doesn't cause emergent to crash