Exclude "savetmp" files from checking in as .dat files
If you save out files every epoch, there is a chance that you are just replacing the files while the cluster_run script scans for dats.
It then tries to add the temp files, which then might fail because the file gets delted while still trying to finish off the svn checkin.