Move SVN update for ClusterRun into a background thread.
When you hit update in ClusterRun, this now spawns a background thread, that goes out and does an svn update
on the local svn working directory. Once that svn update has finished, it signals the main gui thread to then
reload all of the clusterrun tables from local disk.

It also issues a UrlExists on the repository in the gui thread first, to ensure that svn has the neccessary credentials
to access the repositories, as opening a password dialog in the background thread does not work.

This commit only moves the update into the background thread, i.e. all write operations, including the write of "Update Running"
still remain syncronous. Also the initial check for access in the first time hitting "Update" after emergent starts will be syncronous.

So this commit does not eliminate all beachballing in ClusterRun, but should help the most significant and noticeable ones.

As updates now happen in the background, if the "Update Running" stops beachballing, it does not mean the operation is complete
and the data is available for "Load Data". It only means it has written the command to run "Update running" on the backend into the
repository. To see when the update running is actually complete, one needs to look at the SVN cur_svn_rev number of when the upgrade succeeded.