Run an svn cleanup when we start the cluster run server script.
For one, it might be that we need to restart the cluster run script, because it previously crashed and left the svn repository in a locked state

However, the cluster_run svn directory can also accululate a huge amount of cruft in the .svn directory that we need to periodically cleanup.

The .svn directory appears to have a complete copy of everything. But unlike the main version, if one deletes files with svn rm, they don't seem
to get deleted from the .svn copy until one runs a svn cleanup. This led to e.g on cluster run directory shrinking from 70GB to 3 GB after running an svn cleanup