Fix recording of act_g.
The restriction of calculating times.current_cycle_gate_cycle only in thread 0 was not valid,
as the GPi units might not be processed in thread 0 at all.

Also, as we are only setting this value in multiple threads, and all threads should be setting it to the same value,
there shouldn't be any issues with race conditions for this variable, even if multiple threads end up setting it.