Provide an option to disable cluster run cluster_info in backend script
When running on large public clusters, the queues often contain hundreds if
not thousands of jobs concurrently. Polling each to fill the cluster_info
table at that point becomes prohibitively expensive, so disable recording
of the cluster info table.

Also expose the use_qos flag in the main script to be able to switch
between specifying qos or partitions on slurm