Add a function that allows to automatically define the cluster run repository and cluster for the public XSEDE cluster commet at SDSC.
The computing resources are provided by the Neuroscience Gateway portal ( https://www.nsgportal.org/ ) under an NSF grant.

The resources are available for people to more easily perform research in computational neuroscience and Emergent is hopefully
going to be one of the computational neuroscience tools they offer and support.

By including the configuration of this cluster in emergent, this allows to simplify the setup of its usage for new users.

The needed repository for the cluster_run is hosted on grey (clusterun_public), and users need to obtain an svn username and password
in order to ensure accountability for computing resources, by emailing.

More information about this public resource will become available under https://grey.colorado.edu/emergent/index.php/NSG_public