Add an additional pure label column to the cluster run tables.
The previous label column also added the svn tag to the label.

When running paralele batch, the svn tag is different for all currently running batches,
however once completed, they get combined into a single tag.

When the runs are complete, the EpochCompare program can do a straightforward grouping on the
label to create batch averages for the runs. But while the runs are still on going, the labels
are different for each batch in the run and thus averaging doesn't work. By adding a pure label
column, it should be possible to create batch averages both while still running and once completed