Add the ability to block the ClusterRun::Run_impl until the cluster script picks up the jobs
At the moment the ClusterRun just writes into the jobs_submit table overwriting what is there if the cluster run script on the server
hasn't picked it up yet. So if you try and launch a bunch of jobs from a script, some of the jobs might not launch.
So instead add an option to block the function until the jobs_submit  table is empty again